requirements.txt
python-telegram-bot[job-queue]
httpx
python-dotenv
beautifulsoup4
lxml
